backend/requirements/development.txt
# Include base requirements
-r base.txt

# Development and debugging tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Documentation
Sphinx==7.2.6

# Development server enhancements
watchdog==3.0.0

# AI Development Tools
jupyter>=1.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.0

# 3D Mesh Development Tools
meshio>=5.3.0
trimesh>=4.0.0 python-dotenv
